pyproject.toml
[project]
name = "LLM-study"
version = "0.1.0"
description = "LLM study with GitHub integration and LangChain agents"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.11",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.14",
    "langchain-aws>=0.2.24",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.2.0",
    "PyGithub>=2.1.1",
    "python-dotenv>=1.0.0",
    "asyncio>=3.4.3",
    "typing-extensions>=4.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.3",
    "mypy>=1.15.0",
    "pre-commit>=4.0.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
